requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.7.0
Pillow>=9.5.0

# Deep learning models
open-clip-torch>=2.20.0
git+https://github.com/ChaoningZhang/MobileSAM.git

# Configuration management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Visualization
rerun-sdk>=0.17.0

# 3D processing
open3d>=0.17.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
scipy>=1.10.0
scikit-learn>=1.2.0

# FAISS for similarity search (CPU version)
faiss-cpu>=1.7.4

# Segmentation and detection
ultralytics>=8.0.0
supervision>=0.20.0

# Additional utilities
natsort>=8.0.0
imageio>=2.0.0

# Optional: for record3d support (can install later if needed)
# pyliblzfse
